REFACTOR: blueprints for routes, middleware, model, etc
