+++ backend/auth/__init__.py
@@ -1,0 +1,4 @@
+from main import app
+from auth.routes import auth_bp
+
+app.register_blueprint(auth_bp)
