+++ backend/chats/__init__.py
@@ -1,0 +1,4 @@
+from main import app
+from chats.routes import chats_bp
+
+app.register_blueprint(chats_bp)
