+++ backend/location/__init__.py
@@ -1,0 +1,5 @@
+from main import app
+from location.routes import location_bp
+
+
+app.register_blueprint(location_bp)
